added a soft-refresh with useNavigate in the modal component in order to re-trigger the fetch of the user's updated profile pic
